refactor: move code to utils.py
