+++ tests/trainers/decoder_trainer_test.py
@@ -18,6 +18,7 @@ from trainite.config import (
 )
 from trainite.datasets.string_reverse import DatapointModel
 from trainite.trainers.decoder_trainer import Trainer, TrainerConfig, ProjectConfig
+from trainite.shared.utils import upload_model_to_wandb
 from ignite.engine import Events
 from ignite.handlers import EarlyStopping
 import ignite.distributed as idist
@@ -387,7 +388,12 @@ def test_upload_model_to_wandb(project_config, temp_run_dir):
     trainer.exp_logger = mock.MagicMock()
     trainer.checkpointers = {"checkpoint_best": mock.MagicMock(last_checkpoint="best_model_1.pt")}
 
-    trainer._upload_model_to_wandb()
+    upload_model_to_wandb(
+        trainer.exp_logger,
+        trainer.checkpointers,
+        trainer.config.output.run_name,
+        trainer.logger,
+    )
 
     trainer.exp_logger.Artifact.return_value.add_file.assert_called_once_with("best_model_1.pt")
     trainer.exp_logger.log_artifact.assert_called_once_with(trainer.exp_logger.Artifact.return_value)
@@ -399,7 +405,12 @@ def test_upload_model_to_wandb_no_checkpoint(project_config):
     trainer.checkpointers = {}  # nothing saved -> skip, don't crash
 
     with mock.patch.object(trainer.logger, "warning") as mock_warning:
-        trainer._upload_model_to_wandb()
+        upload_model_to_wandb(
+            trainer.exp_logger,
+            trainer.checkpointers,
+            trainer.config.output.run_name,
+            trainer.logger,
+        )
 
     trainer.exp_logger.log_artifact.assert_not_called()
     mock_warning.assert_called_once()
