+++ trainite/shared/utils.py
@@ -1,10 +1,35 @@
+from datetime import datetime
 import importlib
+import inspect
+import logging
 from pathlib import Path
-from typing import Any, TypeVar
+from typing import Any, Literal, TypeVar
 
-import yaml
+from ignite.engine import Engine, Events
+from ignite.handlers import (
+    Checkpoint,
+    DiskSaver,
+    EarlyStopping,
+    create_lr_scheduler_with_warmup,
+)
+from ignite.handlers.fbresearch_logger import FBResearchLogger
+from ignite.handlers.param_scheduler import ParamScheduler
+from ignite.handlers.tensorboard_logger import TensorboardLogger
+from ignite.handlers.wandb_logger import WandBLogger
+from ignite.utils import setup_logger
 from omegaconf import OmegaConf
 from pydantic import BaseModel
+import torch
+from torch import nn
+from torch.optim.lr_scheduler import LinearLR
+from torch.utils.data import DataLoader, Dataset, random_split
+import yaml
+
+from trainite.config.base import (
+    DataConfigBase,
+    DataWithAutoSplit,
+)
+from trainite.datasets.transformed import TransformedDataset
 
 T = TypeVar("T", bound=BaseModel)
 
@@ -72,3 +97,301 @@ def dump_config(config: BaseModel, path: str | Path) -> None:
 def load_config(path: str | Path, config_cls: type[T]) -> T:
     raw_conf = OmegaConf.load(path)
     return config_cls.model_validate(raw_conf)
+
+
+# ==========================================
+# Builders (Dataset, Dataloader, Model Setup)
+# ==========================================
+
+
+# Inspects the target symbol's signature and filters the candidates to
+# only include those that are accepted by the target symbol.
+def _inject_if_accepted(target_symbol: Any, **candidates: Any) -> dict[str, Any]:
+    try:
+        sig = inspect.signature(target_symbol)
+        return {k: v for k, v in candidates.items() if k in sig.parameters}
+    except Exception:
+        return {}
+
+
+# Builds the model based on the provided configuration, tokenizer, and device.
+def build_model(model_config: Any, tokenizer: Any, vocab_size: int, device: str | torch.device) -> nn.Module:
+    target_symbol = get_target(model_config.target)
+    kwargs = _inject_if_accepted(
+        target_symbol,
+        vocab_size=vocab_size,
+        pad_token_id=tokenizer.pad_token_id,
+    )
+    model = instantiate(model_config, **kwargs)
+    model.to(device)
+    return model
+
+
+def build_dataset(dataset_config: Any, transform_config: Any, tokenizer: Any) -> Dataset:
+    ds = get_target(dataset_config.target)
+    dataset = instantiate(dataset_config, **_inject_if_accepted(ds, preprocessor=tokenizer, tokenizer=tokenizer))
+    transform = None
+    if transform_config is not None:
+        tf = get_target(transform_config.target)
+        transform = instantiate(
+            transform_config, **_inject_if_accepted(tf, preprocessor=tokenizer, tokenizer=tokenizer)
+        )
+    # Wraps the dataset with the transform if provided, otherwise returns the dataset as is.
+    return TransformedDataset(dataset, transform)
+
+
+def create_dataloader(
+    dataset: Dataset,
+    dl_config: Any,
+    tokenizer: Any,
+    shuffle: bool | None = None,
+) -> DataLoader:
+    dl_kwargs = dl_config.model_dump(exclude={"collate_fn", "shuffle"})
+    if shuffle is None:
+        shuffle = getattr(dl_config, "shuffle", False)
+    collate_fn = None
+    collate_config = dl_config.collate_fn
+    if collate_config:
+        target_symbol = get_target(collate_config.target)
+        if isinstance(target_symbol, type):
+            collate_fn = instantiate(collate_config, tokenizer=tokenizer)
+        else:
+            collate_fn = target_symbol
+    return DataLoader(dataset, shuffle=shuffle, collate_fn=collate_fn, **dl_kwargs)
+
+
+def _loaders_from_splits(
+    data_config: DataConfigBase, tokenizer: Any
+) -> tuple[DataLoader, DataLoader, DataLoader | None]:
+    def _make(split_config: Any) -> DataLoader:
+        ds = build_dataset(split_config.dataset, split_config.transform, tokenizer)
+        return create_dataloader(ds, split_config.dataloader, tokenizer)
+
+    return _make(data_config.train), _make(data_config.val), _make(data_config.test) if data_config.test else None
+
+
+def _loaders_from_ratios(
+    data_config: DataWithAutoSplit, tokenizer: Any, seed: int
+) -> tuple[DataLoader, DataLoader, DataLoader | None]:
+    dataset = build_dataset(data_config.dataset, data_config.transform, tokenizer)
+    total_len = len(dataset)  # type: ignore
+    if total_len == 0:
+        raise ValueError("Training dataset is empty. Cannot perform train/val/test split.")
+    val_ratio = data_config.val_ratio
+    test_ratio = data_config.test_ratio
+    train_len = int(total_len * (1.0 - test_ratio - val_ratio))
+    val_len = int(total_len * val_ratio)
+    test_len = total_len - train_len - val_len
+    train_ds, val_ds, test_ds = random_split(
+        dataset,
+        [train_len, val_len, test_len],
+        generator=torch.Generator().manual_seed(seed),
+    )
+    dl = data_config.dataloader
+    return (
+        create_dataloader(train_ds, dl, tokenizer, shuffle=True),
+        create_dataloader(val_ds, dl, tokenizer, shuffle=False),
+        create_dataloader(test_ds, dl, tokenizer, shuffle=False) if test_len > 0 else None,
+    )
+
+
+def build_dataloaders(data_config: Any, tokenizer: Any, seed: int) -> tuple[DataLoader, DataLoader, DataLoader | None]:
+    if isinstance(data_config, DataWithAutoSplit):
+        return _loaders_from_ratios(data_config, tokenizer, seed)
+    return _loaders_from_splits(data_config, tokenizer)
+
+
+# ==========================================
+# Handlers (Ignite Engine Wiring & Setup)
+# ==========================================
+
+
+def make_run_dir(output_config: Any) -> Path:
+    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
+    run_name = output_config.run_name
+    run_dir = Path(output_config.root) / run_name / timestamp
+    run_dir.mkdir(parents=True, exist_ok=True)
+    return run_dir
+
+
+def attach_lr_scheduler(
+    engine: Engine,
+    optimizer: Any,
+    total_iters: int,
+    peak_lr: float,
+) -> ParamScheduler:
+    warmup_iters = max(2, int(0.1 * total_iters))
+    linear_decay = LinearLR(
+        optimizer,
+        start_factor=1.0,
+        end_factor=0.0,
+        total_iters=total_iters - warmup_iters,
+    )
+    scheduler: ParamScheduler = create_lr_scheduler_with_warmup(
+        linear_decay,
+        warmup_start_value=0.0,
+        warmup_end_value=peak_lr,
+        warmup_duration=warmup_iters,
+    )
+    engine.add_event_handler(Events.ITERATION_COMPLETED, scheduler)
+    return scheduler
+
+
+def attach_early_stopping(
+    val_evaluator: Engine,
+    trainer_engine: Engine,
+    patience: int | None,
+) -> EarlyStopping | None:
+    if patience is not None:
+        early_stopping = EarlyStopping(
+            patience=patience,
+            score_function=lambda engine: engine.state.metrics["loss"],
+            trainer=trainer_engine,
+            min_delta=0.0,
+            mode="min",
+        )
+        val_evaluator.add_event_handler(Events.COMPLETED, early_stopping)
+        return early_stopping
+    return None
+
+
+def setup_checkpointing(
+    engine: Engine,
+    val_evaluator: Engine,
+    to_save: dict[str, Any],
+    run_dir: Path,
+) -> dict[str, Checkpoint]:
+    checkpointers = {}
+
+    def score_function(engine_val):
+        loss = engine_val.state.metrics["loss"]
+        return -loss
+
+    checkpoint = Checkpoint(
+        to_save=to_save,
+        save_handler=DiskSaver(dirname=str(run_dir), require_empty=False),
+        filename_prefix="best",
+        score_function=score_function,
+        score_name="val_loss",
+        n_saved=1,
+        global_step_transform=lambda *_: engine.state.iteration,
+    )
+    val_evaluator.add_event_handler(Events.COMPLETED, checkpoint)
+    checkpointers["checkpoint_best"] = checkpoint
+
+    last_checkpoint = Checkpoint(
+        to_save=to_save,
+        save_handler=DiskSaver(dirname=str(run_dir), require_empty=False),
+        filename_prefix="last",
+        n_saved=1,
+        global_step_transform=lambda *_: engine.state.iteration,
+    )
+    engine.add_event_handler(Events.EPOCH_COMPLETED, last_checkpoint)
+
+    checkpointers["checkpoint_last"] = last_checkpoint
+    return checkpointers
+
+
+def setup_experiment_logger(
+    backend: Literal["tensorboard", "wandb"],
+    engine: Engine,
+    val_evaluator: Engine,
+    train_evaluator: Engine,
+    test_evaluator: Engine,
+    optimizer: Any,
+    run_dir: Path,
+    metric_names: list[str],
+    has_test: bool,
+    run_name: str,
+) -> TensorboardLogger | WandBLogger:
+    if backend == "wandb":
+        exp_logger: TensorboardLogger | WandBLogger = WandBLogger(
+            project=run_name, dir=str(run_dir), name=str(run_dir).split("/")[-1]
+        )
+    else:
+        log_dir = run_dir / "tensorboard" if run_dir else None
+        exp_logger = TensorboardLogger(log_dir=log_dir)
+
+    # Log training iteration loss
+    exp_logger.attach_output_handler(
+        engine,
+        event_name=Events.ITERATION_COMPLETED,
+        tag="training",
+        output_transform=lambda output: {"batch_loss": output["loss"]},
+    )
+
+    # Log training epoch metrics
+    exp_logger.attach_output_handler(
+        train_evaluator,
+        event_name=Events.EPOCH_COMPLETED,
+        tag="training",
+        metric_names=metric_names,
+        global_step_transform=lambda *_: engine.state.iteration,
+    )
+
+    # Log validation epoch metrics
+    exp_logger.attach_output_handler(
+        val_evaluator,
+        event_name=Events.EPOCH_COMPLETED,
+        tag="validation",
+        metric_names=metric_names,
+        global_step_transform=lambda *_: engine.state.iteration,
+    )
+
+    # Log test metrics if applicable
+    if has_test:
+        exp_logger.attach_output_handler(
+            test_evaluator,
+            event_name=Events.COMPLETED,
+            tag="testing",
+            metric_names=metric_names,
+            global_step_transform=lambda *_: engine.state.iteration,
+        )
+
+    # Log optimizer learning rates
+    exp_logger.attach_opt_params_handler(
+        engine,
+        event_name=Events.ITERATION_STARTED,
+        optimizer=optimizer,
+    )
+    return exp_logger
+
+
+def upload_model_to_wandb(
+    exp_logger: Any,
+    checkpointers: dict[str, Checkpoint],
+    run_name: str,
+    logger: logging.Logger,
+) -> None:
+    checkpoint = checkpointers.get("checkpoint_best") or checkpointers.get("checkpoint_last")
+    checkpoint_path = checkpoint.last_checkpoint if checkpoint else None
+    if not checkpoint_path:
+        logger.warning("No checkpoint found; skipping model upload to Weights & Biases.")
+        return
+    artifact = exp_logger.Artifact(name=f"{run_name}-model".replace("/", "-"), type="model")
+    artifact.add_file(str(checkpoint_path))
+    exp_logger.log_artifact(artifact)
+    logger.info("Uploaded model checkpoint to Weights & Biases: %s", checkpoint_path)
+
+
+def setup_console_logger(
+    run_dir: Path,
+    engine: Engine,
+    log_every_steps: int,
+    optimizer: Any,
+) -> logging.Logger:
+    logger = setup_logger(
+        "trainer",
+        level=logging.INFO,
+        filepath=str(run_dir / "output.log") if run_dir else None,
+        reset=True,
+    )
+    train_fb_logger = FBResearchLogger(logger=logger, show_output=True)
+    train_fb_logger.attach(
+        engine,
+        name="Train",
+        every=log_every_steps,
+        optimizer=optimizer,
+        output_transform=lambda output: {"loss": output["loss"].item()},
+    )
+    return logger
